1079 refactor our tests & service to fit ofbiz Double to BigDecimal change
git-svn-id: svn://svn.opentaps.org/opentaps_all/branches/ofbiz-0904-1.4.0-preview1@13090 d3523486-f5fe-0310-a41a-a15a4e76f3c7

+++ hot-deploy/financials/src/com/opensourcestrategies/financials/util/UtilCOGS.java
@@ -317,7 +317,7 @@ public final class UtilCOGS {
                 if (ServiceUtil.isError(serviceResults)) {
                     throw new GenericServiceException(ServiceUtil.getErrorMessage(serviceResults));
                 }
-                BigDecimal facilityQuantity = BigDecimal.valueOf(((Double) serviceResults.get("quantityOnHandTotal")).doubleValue());
+                BigDecimal facilityQuantity = (BigDecimal) serviceResults.get("quantityOnHandTotal");
                 quantity = quantity.add(facilityQuantity).setScale(decimals, rounding); // TODO: quantity sum needs its own decimals and rounding
             }
         } else {
